Now using latest Presentation API polyfill
This is a first attempt to integrate the Physical Web prototype with the
latest version of the Presentation API polyfill, which also supports Chromecast
devices and opening the presentation in a separate window.

The code is untested (and thus wrong!) since uri-beacon is Lunix-only and
I don't have the right environment here.